detcore

Module preemptions

A datatype to abstract a record of thread preemptions, as generated during chaos mode execution.

Structs

PreemptionReader

A reader for a stream of preemption events.

PreemptionRecord

A record of all the preemptions and other scheduling events that occur during execution.

PreemptionWriter

A writer for a stream of preemption events.

ThreadHistory

The record of priorities and preemptions for a particular thread.

ThreadHistoryIterator

An iterator over the preemption history of a thread. Does not include the final timeslice, which does not have an ending time.

Functions

read_trace

Read a full trace from disk. Panic if it doesn’t load.

strip_times_from_events_file

Take a file containing a PreemptionRecord and strip the times from all the events. Optionally take a destination path, otherwise create a destination file in the same directory with our own naming convention.