detcore

Module types

Widely-shared type definitions.

Structs

DetPid

Deterministic “virtual” version of reverie::Pid

DetTime

A determinstic notion of time, measuring progress of one thread.

FdSlot

Identity of one numeric descriptor slot within a descriptor table.

FilesId

Identity of a Linux descriptor table (files_struct).

GlobalTime

Deterministic global time, combining local times.

LogicalTime

Represents an absolute point in time in nanoseconds. Parts of this API are largely inspired by std::time::Duration. This could go to 128 bits if we need more than ~585 years of nanosecond precision.

MmId

Identity of a Linux memory address space (mm_struct).

OpenFileId

Identity of a Linux open file description (struct file).

SchedEvent

A scheduled action by one thread in the system. This can be recorded, or replayed to guide the schedule.

SigWrapper

Simply a type to hang Serialize/Deserialize instances off of.

Enums

FutexID

Identity of a futex word.

Op

NOTE [Event Semantics]

SharedMemoryObjectId

Identity of an object that backs a process-shared memory mapping.

SyscallPhase

Which phase of the syscall did we observe on a given event: the prehook or the posthook.

Constants

NANOS_PER_NONDET_INSTR

Virtual nanoseconds elapsed per nondeterministic instruction other than system calls.

NANOS_PER_RCB

Virtual nanoseconds elapsed per Retired Conditional Branch.

NANOS_PER_SCHED

Virtual nanoseconds elapsed per step of the scheduler.

NANOS_PER_SYSCALL

Virtual nanoseconds elapsed per system call (uniform for now).

Type Aliases

DetInode

Deterministic “virtual” inode.

DetTid

Deterministic “virtual” version of reverie::Tid

InstructionPointer

The type of the RIP value.

LogicalDuration

Microseconds

The same basic type alias as nanoseconds. Just for clarity/readability.

RawFd

For now we use the definiton of RawFd from std::os.

RawInode

Nondeterministic “physical” inode