detcore

Module logdiff

Everything to do with post-processing hermit/detcore logs.

Structs

Comparison

A comparison of two strings.

LogDiffOpts

Options for calling log_diff.

Enums

DetLogFilter

Indicates which DETLOG entries to be used for log-diff comparison

LogComparisonMode

Selects the set of log messages compared for determinism.

Functions

log_diff

Process log messages from two files. Log messages look like this: “Apr 09 06:08:03.100 INFO detcore: [detcore, dtid 2] finish syscall: close(2) = Ok(0)”

strip_log_entry

In fully-deterministic modes, many log lines should be fully determinstic across runs. But as that is a work-in-progress, this utility strips known-nondeterministic information from logs.