detcore::logdiff

Function log_diff

pub fn log_diff(file_a: &Path, file_b: &Path, opts: &LogDiffOpts) -> bool

Process log messages from two files. Log messages look like this: “Apr 09 06:08:03.100 INFO detcore: [detcore, dtid 2] finish syscall: close(2) = Ok(0)”

With some complexities:

Some entries are multi-line (contain newlines).

Some stripping of nondeterministic information is needed for direct comparability.

Certain lines are intended to be deterministic/comparable, in their contents, and others in their presence but not their details.